reorganize packages; remove unecesarry interace declarations
