+++ bitcoinspv/clients/btcwrapper/query.go
