--- btcwrapper/interface.go
@@ -1,26 +1,0 @@
-package btcwrapper
-
-import (
-	"github.com/btcsuite/btcd/chaincfg/chainhash"
-	btcwire "github.com/btcsuite/btcd/wire"
-
-	relayertypes "github.com/gonative-cc/relayer/bitcoinspv/types"
-	btctypes "github.com/gonative-cc/relayer/bitcoinspv/types/btc"
-)
-
-// BTCClient defines the interface for interacting with a Bitcoin node
-type BTCClient interface {
-	// Node lifecycle methods
-	Stop()
-	WaitForShutdown()
-
-	// Block subscription methods
-	SubscribeNewBlocks()
-	BlockEventChannel() <-chan *btctypes.BlockEvent
-
-	// Block query methods
-	GetBTCTipBlock() (*chainhash.Hash, int64, error)
-	GetBTCBlockByHash(blockHash *chainhash.Hash) (*relayertypes.IndexedBlock, *btcwire.MsgBlock, error)
-	GetBTCTailBlocksByHeight(Blockheight int64) ([]*relayertypes.IndexedBlock, error)
-	GetBTCBlockByHeight(Blockheight int64) (*relayertypes.IndexedBlock, *btcwire.MsgBlock, error)
-}
